Factor out components for xref display
