+++ ui/src/components/XrefTable/XrefTableRow.jsx
@@ -1,10 +1,11 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { injectIntl, FormattedNumber } from 'react-intl';
+import { FormattedNumber } from 'react-intl';
 
 import {
-  Collection, Entity, Property, Skeleton, JudgementButtons,
+  Collection, Skeleton, JudgementButtons,
 } from 'components/common';
+import EntityCompare from 'components/Entity/EntityCompare';
 import { showWarningToast } from 'app/toast';
 import { pairwiseJudgement } from 'actions';
 
@@ -43,43 +44,6 @@ class XrefTableRow extends Component {
     );
   }
 
-  getCommonProperties() {
-    const { entity, match } = this.props.xref;
-    const properties = [...entity.schema.getFeaturedProperties()];
-
-    match.schema.getFeaturedProperties().forEach((prop) => {
-      if (properties.indexOf(prop) === -1) {
-        properties.push(prop);
-      }
-    });
-
-    return properties.filter((prop) => {
-      return entity.hasProperty(prop) || match.hasProperty(prop);
-    });
-  }
-
-  renderProperties(entity) {
-    const properties = this.getCommonProperties();
-
-    return (
-      <>
-        <Entity.Link entity={entity} preview icon />
-        <div className="XrefTableRow__properties">
-          {properties.map((prop) => (
-            <div className="XrefTableRow__property" key={prop.name}>
-              <span className="XrefTableRow__property__name text-muted">
-                <Property.Name prop={prop} />
-              </span>
-              <span className="XrefTableRow__property__value">
-                <Property.Values prop={prop} values={entity.getProperty(prop)} translitLookup={entity.latinized} />
-              </span>
-            </div>
-          ))}
-        </div>
-      </>
-    );
-  }
-
   render() {
     const { isPending, xref } = this.props;
 
@@ -96,10 +60,10 @@ class XrefTableRow extends Component {
           <JudgementButtons obj={xref} onChange={this.onDecide} />
         </td>
         <td className="entity bordered">
-          {this.renderProperties(xref.entity)}
+          <EntityCompare entity={xref.entity} other={xref.match} />
         </td>
         <td className="entity">
-          {this.renderProperties(xref.match)}
+          <EntityCompare entity={xref.match} other={xref.entity} />
         </td>
         <td className="numeric narrow">
           <FormattedNumber value={parseInt(parseFloat(xref.score) * 100, 10)} />
@@ -113,4 +77,4 @@ class XrefTableRow extends Component {
 }
 
 XrefTableRow = connect(null, { pairwiseJudgement })(XrefTableRow);
-export default injectIntl(XrefTableRow);
+export default XrefTableRow;
